+++ jooby/src/test/java/org/jooby/internal/JvmInfoTest.java
@@ -15,7 +15,6 @@
  */
 package org.jooby.internal;
 
-import static org.easymock.EasyMock.expect;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
@@ -23,12 +22,7 @@ import java.lang.management.ManagementFactory;
 
 import org.jooby.test.MockUnit;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.powermock.core.classloader.annotations.PrepareForTest;
-import org.powermock.modules.junit4.PowerMockRunner;
 
-@RunWith(PowerMockRunner.class)
-@PrepareForTest({JvmInfo.class, ManagementFactory.class })
 public class JvmInfoTest {
 
   @Test
@@ -45,8 +39,7 @@ public class JvmInfoTest {
   public void piderr() throws Exception {
     new MockUnit()
     .expect(unit -> {
-      unit.mockStatic(ManagementFactory.class);
-      expect(ManagementFactory.getRuntimeMXBean()).andThrow(new RuntimeException());
+      unit.mockStatic(ManagementFactory.class).when(ManagementFactory::getRuntimeMXBean).thenThrow(new RuntimeException());
     })
     .run(unit -> {
       assertEquals(-1, JvmInfo.pid());