+++ jooby/src/test/java/org/jooby/internal/mvc/RequestParamNameProviderTest.java
@@ -15,7 +15,7 @@
  */
 package org.jooby.internal.mvc;
 
-import static org.easymock.EasyMock.expect;
+import static org.mockito.Mockito.when;
 import static org.junit.Assert.assertEquals;
 
 import java.lang.reflect.Method;
@@ -25,12 +25,7 @@ import org.jooby.Env;
 import org.jooby.internal.RouteMetadata;
 import org.jooby.test.MockUnit;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.powermock.core.classloader.annotations.PrepareForTest;
-import org.powermock.modules.junit4.PowerMockRunner;
 
-@RunWith(PowerMockRunner.class)
-@PrepareForTest(RequestParam.class  )
 public class RequestParamNameProviderTest {
 
   public void dummy(final String dummyparam) {
@@ -44,11 +39,10 @@ public class RequestParamNameProviderTest {
     new MockUnit(Env.class)
         .expect(unit -> {
           Env env = unit.get(Env.class);
-          expect(env.name()).andReturn("dev");
+          when(env.name()).thenReturn("dev");
         })
         .expect(unit -> {
-          unit.mockStatic(RequestParam.class);
-          expect(RequestParam.nameFor(param)).andReturn(null);
+          unit.mockStatic(RequestParam.class).when(() -> RequestParam.nameFor(param)).thenReturn(null);
         })
         .run(unit -> {
           assertEquals("dummyparam",